extract shared pipeline date helpers + derive meeting-type from MeetingInput
Applies the audit findings from /ts-audit:

- Move extractMeetingDateFromTitle, normalizeEgovDate, and
  normalizeFinalsiteDate into src/pipeline/dates.ts. Both ScraperService
  and the orchestrator had their own copies of the month-name lookup,
  and both typed it as Record<string, string> which widened the value
  union from literal months to plain strings. The shared table uses
  'as const satisfies Record<string, string>' so keyof and value types
  stay narrow.
- meetingTypeFromFinalsiteLabel's return type is now
  MeetingInput['meetingType'] via indexed access, so a future variant
  stays in lockstep with the storage contract instead of drifting
  against a duplicated string-literal union.

Pure refactor — behavior unchanged, all 136 tests still green.
